requirements.txt
cryptography>=41.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
